refactor: 상세 화면 UI 상태 관련 모델 패키지 구조 개선
- `AfternoteDetailUiState` 및 `AfternoteDeleteState`를 `detail.model` 서브 패키지에서 `detail` 패키지로 이동하여 구조 단순화
- 패키지 이동에 따른 상세 화면 관련 컴포저블(Gallery, MemorialGuideline, SocialNetwork) 및 ViewModel의 임포트 경로 수정
- `AfternoteNavGraphDetail` 및 `AfternoteDetailViewModel` 내 상태 처리 로직에 중괄호를 추가하여 코드 가독성 및 일관성 개선
