+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/detail/AfternoteDetailUiState.kt
@@ -1,6 +1,4 @@
-package com.afternote.feature.afternote.presentation.author.detail.model
-
-import com.afternote.feature.afternote.presentation.author.detail.DetailContentUiModel
+package com.afternote.feature.afternote.presentation.author.detail
 
 /**
  * 애프터노트 상세 화면 UI 상태.
