+++ feature/afternote/presentation/src/main/kotlin/com/afternote/feature/afternote/presentation/author/navigation/AfternoteNavGraphDetail.kt
@@ -16,10 +16,10 @@ import androidx.navigation.NavBackStackEntry
 import androidx.navigation.toRoute
 import com.afternote.core.ui.topbar.DetailTopBar
 import com.afternote.feature.afternote.presentation.R
+import com.afternote.feature.afternote.presentation.author.detail.AfternoteDeleteState
 import com.afternote.feature.afternote.presentation.author.detail.GalleryDetailRoute
 import com.afternote.feature.afternote.presentation.author.detail.MemorialGuidelineDetailRoute
 import com.afternote.feature.afternote.presentation.author.detail.SocialNetworkDetailRoute
-import com.afternote.feature.afternote.presentation.author.detail.model.AfternoteDeleteState
 import com.afternote.feature.afternote.presentation.author.navigation.model.AfternoteRoute
 
 @Composable
@@ -69,13 +69,17 @@ internal fun HandleDeleteResult(
                 onBack()
                 onConsumed()
             }
+
             is AfternoteDeleteState.Failed -> {
                 // 에러 UI 처리는 화면별 Snackbar/Dialog 에서 담당. 여기서는 상태만 소비.
                 onConsumed()
             }
+
             AfternoteDeleteState.Idle,
             AfternoteDeleteState.InProgress,
-            -> Unit
+            -> {
+                Unit
+            }
         }
     }
 }
